+++ js/data.js
@@ -1,0 +1,65 @@
+const NUMS_FOR_IMAGE_URL = ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10'];
+
+const CHARS_FOR_PHOTO_URL = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+
+const TITLES = [
+  'Петровский Арт Лофт',
+  'Agent Flat',
+  'Гранд',
+  'Avenue-Apart',
+  'ReStop Loft Apart',
+  'Greenfeel',
+  'Smart Apart Africa',
+  'Lifehack Apart',
+  'Loft Apart',
+  'Англетер',
+];
+
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+const CHECKINS = ['12:00', '13:00', '14:00'];
+
+const CHECKOUTS = ['12:00', '13:00', '14:00'];
+
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+const DESCRIPTIONS = [
+  'Все номера оснащены телевизором с плоским экраном и спутниковыми каналами.',
+  'Все апартаменты оснащены телевизором с плоским экраном, микроволновой печью, холодильником и чайником.',
+  'Парам особенно нравится расположение — они оценили проживание в этом районе.',
+  'Гости могут готовить на полностью оборудованной мини-кухне.',
+  'Во всех номерах есть мини-кухня, телевизор с плоским экраном и кабельными каналами.',
+  'В числе удобств телевизор с кабельными каналами.',
+  'В распоряжении гостей холодильник, плита и чайник.',
+  'В распоряжении гостей апартаменты с гамаком, обеденной зоной, гостиным уголком.',
+  'Все номера гостевого дома оснащены чайником. Кроме того, в распоряжении гостей микроволновая печь.',
+  'Зарядитесь энергией в тренажерном зале. Восстановите силы в сауне.',
+];
+
+const MIN_PRICE = 2000;
+const MAX_PRICE = 20000;
+const MIN_ROOMS = 1;
+const MAX_ROOMS = 5;
+const MIN_GUESTS = 1;
+const MAX_GUESTS = 10;
+const LATITUDE = [35.65, 35.7];
+const LONGITUDE = [139.7, 139.8];
+
+export {
+  NUMS_FOR_IMAGE_URL,
+  CHARS_FOR_PHOTO_URL,
+  TITLES,
+  TYPES,
+  CHECKINS,
+  CHECKOUTS,
+  FEATURES,
+  DESCRIPTIONS,
+  MIN_PRICE,
+  MAX_PRICE,
+  MIN_ROOMS,
+  MAX_ROOMS,
+  MIN_GUESTS,
+  MAX_GUESTS,
+  LATITUDE,
+  LONGITUDE
+};
